test 7

+++ data.c
@@ -3,7 +3,7 @@
 
 struct test{
   char first;
-  long second;
+  float second;
   int third;
   //char fourth;
   //char* fifth;
@@ -13,6 +13,6 @@ int main(int argc, char *argv[])
 {
   int size = sizeof(struct test);
   printf("Size is: %d \n", size);
-  printf("Size is: %d \n", (int)sizeof(long));
+  printf("Size is: %d \n", (int)sizeof(float));
   return 0;
 }
